+++ application/blueprints/__init__.py
@@ -1,0 +1,2 @@
+from .temperature import temperature
+from .version import version
